charge types test
